Work on an app settings screen
Work on garden20/garden#12

This allows a user to use a json schema on the application manifest
that defines the form properties. When a user saves this, it will be
saved to the design_doc as app_settings. This allows shows, lists,
views to read this data.

Todo: still need to manage conflicts when the design doc gets updated.