fixed logout bug when require_valid_user == true in couchdb config
logout simply didn't work when require_valid_user was true. i think this
is somewhat by design because /_session does not require auth and
returns 200 no matter what, so it's impossible to delete the session? see
http://wiki.apache.org/couchdb/Session_API  so I repeated a hack which
appears to work in Futon, pass in _:_ for user/pass when doing DELETE
/_session.

Needs more testing with the other login methods, not sure how they
will be affected.