update dashboard core to use app_settings package when available. #89
Only fetches entire ddoc when app_settings paths are not found.

One fetch of a ddoc is still required right now during settings
migration.  The settings migration path is specified in the `kanso` or
`couchapp` property of the ddoc. We should consider extending the
app_settings module to allow for querying abitrary properties on the
ddoc.